+++ CHANGES.rst
@@ -4,5 +4,10 @@ Changelog
 1.0 (unreleased)
 ----------------
 
+- Add ``input_type`` option to be able to better select the type of input.
+  Add ``input_type`` support for dropdowns.
+  Remove ``as_input`` attribute and provide upgrade step for it.
+  [thet]
+
 - Initial release from collective.portlet.collectionfilter.
   [thet]